refactor(charts): hoist pie label parse and trim font types
Parse pie labelcomponent once in VictoryChartPie, import getChartSkiaTypeface
from utils, collapse ChartFontsValue alias, and remove duplicate native provider.
