+++ src/components/Charts/context/ChartFontsContext.tsx
@@ -1,11 +1,11 @@
 import type {SkTypefaceFontProvider} from '@shopify/react-native-skia';
 import {createContext, useContext} from 'react';
-import type {ChartFontsContextValue} from '@components/Charts/types/chartFontsTypes';
+import type {ChartFontsValue} from '@components/Charts/types/chartFontsTypes';
 import type {ChartDefaultTypeface} from '@components/Charts/types/chartSkiaTypefaceTypes';
 
-const ChartFontsContext = createContext<ChartFontsContextValue | null>(null);
+const ChartFontsContext = createContext<ChartFontsValue | null>(null);
 
-function useChartFontsContext(): ChartFontsContextValue {
+function useChartFontsContext(): ChartFontsValue {
     const context = useContext(ChartFontsContext);
 
     if (!context) {
