+++ src/components/HTMLEngineProvider/HTMLRenderers/VictoryChartRenderer/components/VictoryChartPie.tsx
@@ -1,7 +1,10 @@
-import React from 'react';
+import React, {useMemo} from 'react';
+import {useAmbientTRenderEngine} from 'react-native-render-html';
 import type {TNode} from 'react-native-render-html';
 import {Pie} from 'victory-native';
+import parseVictoryLabelNode from '@components/HTMLEngineProvider/HTMLRenderers/VictoryChartRenderer/parsers/victoryLabelParser';
 import parseAttribute from '@components/HTMLEngineProvider/HTMLRenderers/VictoryChartRenderer/utils/parseAttribute';
+import parseComponent from '@components/HTMLEngineProvider/HTMLRenderers/VictoryChartRenderer/utils/parseComponent';
 import VictoryChartPieLabel from './VictoryChartPieLabel';
 
 type VictoryChartPieProps = {tnode: TNode};
@@ -10,6 +13,12 @@ type VictoryChartPieProps = {tnode: TNode};
 const START_ANGLE = 270;
 
 function VictoryChartPie({tnode}: VictoryChartPieProps) {
+    const renderEngine = useAmbientTRenderEngine();
+    const labelItemTemplate = useMemo(() => {
+        const labelComponentNode = parseComponent(tnode.attributes.labelcomponent, renderEngine, 'victorylabel');
+        return labelComponentNode ? parseVictoryLabelNode(labelComponentNode).labelItems?.at(0) : undefined;
+    }, [tnode.attributes.labelcomponent, renderEngine]);
+
     const innerRadius = tnode.attributes.innerradius !== undefined ? Number(parseAttribute(tnode.attributes.innerradius)) : undefined;
     const radius = tnode.attributes.radius !== undefined ? Number(parseAttribute(tnode.attributes.radius)) : undefined;
     const size = radius ? radius * 2 : undefined;
@@ -25,6 +34,7 @@ function VictoryChartPie({tnode}: VictoryChartPieProps) {
                     <VictoryChartPieLabel
                         tnode={tnode}
                         slice={slice}
+                        labelItemTemplate={labelItemTemplate}
                     />
                 </Pie.Slice>
             )}
