+++ src/components/HTMLEngineProvider/HTMLRenderers/VictoryChartRenderer/components/VictoryChartPieLabel.tsx
@@ -1,28 +1,23 @@
 import React from 'react';
-import {useAmbientTRenderEngine} from 'react-native-render-html';
 import type {TNode} from 'react-native-render-html';
 import type {PieSliceData} from 'victory-native';
 import {useVictoryChartContext} from '@components/HTMLEngineProvider/HTMLRenderers/VictoryChartRenderer/context/VictoryChartContext';
-import parseVictoryLabelNode from '@components/HTMLEngineProvider/HTMLRenderers/VictoryChartRenderer/parsers/victoryLabelParser';
-import type {PolarChartData} from '@components/HTMLEngineProvider/HTMLRenderers/VictoryChartRenderer/types';
+import type {LabelItem, PolarChartData} from '@components/HTMLEngineProvider/HTMLRenderers/VictoryChartRenderer/types';
 import parseAttribute from '@components/HTMLEngineProvider/HTMLRenderers/VictoryChartRenderer/utils/parseAttribute';
-import parseComponent from '@components/HTMLEngineProvider/HTMLRenderers/VictoryChartRenderer/utils/parseComponent';
 import VictoryChartLabel from './VictoryChartLabel';
 
 type VictoryChartPieLabelProps = {
     tnode: TNode;
     slice: PieSliceData;
+    labelItemTemplate?: LabelItem;
 };
 
 const RADIAN = Math.PI / 180;
 
-function VictoryChartPieLabel({tnode, slice}: VictoryChartPieLabelProps) {
+function VictoryChartPieLabel({tnode, slice, labelItemTemplate}: VictoryChartPieLabelProps) {
     const {data} = useVictoryChartContext();
-    const renderEngine = useAmbientTRenderEngine();
-    const labelComponentNode = parseComponent(tnode.attributes.labelcomponent, renderEngine, 'victorylabel');
-    const labelItem = labelComponentNode ? parseVictoryLabelNode(labelComponentNode).labelItems?.at(0) : undefined;
 
-    if (!labelItem) {
+    if (!labelItemTemplate) {
         return null;
     }
 
@@ -35,10 +30,13 @@ function VictoryChartPieLabel({tnode, slice}: VictoryChartPieLabelProps) {
     const x = slice.center.x + labelRadius * Math.cos(-midAngle * RADIAN);
     const y = slice.center.y + labelRadius * Math.sin(midAngle * RADIAN);
 
-    labelItem.text = text;
-    labelItem.x = x;
-    labelItem.y = y;
-    labelItem.textAnchor = 'middle';
+    const labelItem: LabelItem = {
+        ...labelItemTemplate,
+        text,
+        x,
+        y,
+        textAnchor: 'middle',
+    };
 
     return <VictoryChartLabel {...labelItem} />;
 }
